backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.2
aiosmtplib==3.0.2
alembic==1.13.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==25.1.0
bcrypt==4.2.1
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.109.0
fastapi-mail==1.4.2
feedparser==6.0.10
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.3
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
-e git+https://github.com/djangify/newsapi.git@fa2aefc7906f193c99c7f46aa8700b5343ea5357#egg=newsapi
passlib==1.7.4
propcache==0.2.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyMySQL==1.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
redis==5.0.1
rsa==4.9
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.12.2
uvicorn==0.27.0
yarl==1.18.3
